# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Erica's Paint & Sip API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all published events
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: upcoming
          in: query
          description: Only return future events
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated list of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
  /events/upcoming:
    get:
      operationId: getUpcomingEvents
      tags: [events]
      summary: Get next N upcoming events (homepage widget)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 3
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /events/summary:
    get:
      operationId: getEventsSummary
      tags: [events]
      summary: Aggregate stats — total events, upcoming count, total bookings (admin only)
      responses:
        "200":
          description: Events summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsSummary"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
  /events/{slug}/by-slug:
    get:
      operationId: getEventBySlug
      tags: [events]
      summary: Get event by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found

  # ── Bookings ─────────────────────────────────────────────────────────────────
...
  /gallery:
    get:
      operationId: listGalleryPhotos
      tags: [gallery]
      summary: List gallery photos
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: Paginated gallery photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhotoListResponse"
    post:
      operationId: createGalleryPhoto
      tags: [gallery]
      summary: Add a gallery photo (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryPhotoBody"
      responses:
        "201":
          description: Created photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
  /gallery/featured:
    get:
      operationId: getFeaturedGalleryPhotos
      tags: [gallery]
      summary: Get featured photos for the homepage (top by sortOrder)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Featured photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryPhoto"
...
  /gallery/{id}:
    get:
      operationId: getGalleryPhoto
      tags: [gallery]
      summary: Get gallery photo by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gallery photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
    put:
      operationId: updateGalleryPhoto
      tags: [gallery]
      summary: Update a gallery photo (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryPhotoBody"
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
    delete:
      operationId: deleteGalleryPhoto
      tags: [gallery]
      summary: Delete a gallery photo (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist

  # ── Blog Posts ───────────────────────────────────────────────────────────────
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to the newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeNewsletterBody"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscriber"
        "409":
          description: Email already subscribed
  /newsletter/subscribers:
    get:
      operationId: listNewsletterSubscribers
      tags: [newsletter]
      summary: List newsletter subscribers (admin only — subscriber PII)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated subscribers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscriberListResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ── Event ──────────────────────────────────────────────────────────────────
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        date:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        capacity:
          type: integer
          nullable: true
        seatsAvailable:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, date, isPublished, createdAt]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        capacity:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        isPublished:
          type: boolean
          default: false
      required: [title, date]
    EventListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    EventsSummary:
      type: object
      properties:
        totalEvents:
          type: integer
        upcomingEvents:
          type: integer
        totalBookings:
          type: integer
        totalRevenue:
          type: number
      required: [totalEvents, upcomingEvents, totalBookings, totalRevenue]

    # ── Booking ────────────────────────────────────────────────────────────────
...
    GalleryPhoto:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        sortOrder:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, imageUrl, createdAt]
    CreateGalleryPhotoBody:
      type: object
      properties:
        imageUrl:
          type: string
        caption:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        sortOrder:
          type: integer
          nullable: true
      required: [imageUrl]
    GalleryPhotoListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GalleryPhoto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]

    # ── Post ───────────────────────────────────────────────────────────────────
...
    NewsletterSubscriber:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        email:
          type: string
        subscribedAt:
          type: string
          format: date-time
      required: [id, email, subscribedAt]
    SubscribeNewsletterBody:
      type: object
      properties:
        name:
          type: string
          nullable: true
        email:
          type: string
      required: [email]
    NewsletterSubscriberListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NewsletterSubscriber"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]